MVCN2TST,TGC: Fix all AVLTree:insertValue to take int64_t
Not int!

This is doubtless responsible for some bugs, but debugging testcase is
still causing assertions - and even earlier on, by the looks of things.

+++ my-own-problems/move-the-coins-2-test/testcase-generator/utils.h
@@ -95,7 +95,7 @@ namespace MVCN2TST
                 else
                     return m_rootForRevision[m_revisionNumber];
             }
-            void insertValue(int newValue)
+            void insertValue(int64_t newValue)
             {
                 if (!m_root)
                     m_root = createNode(newValue);
@@ -119,7 +119,7 @@ namespace MVCN2TST
         private:
             AVLNode* m_root = nullptr;
 
-            AVLNode* insertValue(int newValue, AVLNode* currentNode)
+            AVLNode* insertValue(int64_t newValue, AVLNode* currentNode)
             {
                 if (m_isPersistent)
                 {
@@ -246,7 +246,7 @@ namespace MVCN2TST
                 }
             }
 
-            AVLNode* createNode(int value)
+            AVLNode* createNode(int64_t value)
             {
                 auto newNode = createNode();
                 newNode->value = value;
